delete comment and post done

+++ server/controller/post-controller.js
@@ -17,26 +17,28 @@ export const updatePost = async (request, response) => {
         const post = await Post.findById(request.params.id);
 
         if (!post) {
-            response.status(404).json({ msg: 'Post not found' })
+            return response.status(404).json({ msg: 'Post not found' })
         }
         
         await Post.findByIdAndUpdate( request.params.id, { $set: request.body })
 
-        response.status(200).json('post updated successfully');
+        return response.status(200).json('post updated successfully');
     } catch (error) {
-        response.status(500).json(error);
+        return response.status(500).json(error);
     }
 }
 
 export const deletePost = async (request, response) => {
     try {
         const post = await Post.findById(request.params.id);
-        
+        if(!post){
+            return response.status(404).json({ msg: 'Post not found' })
+        }
         await post.delete()
 
-        response.status(200).json('post deleted successfully');
+        return response.status(200).json('post deleted successfully');
     } catch (error) {
-        response.status(500).json(error)
+        return response.status(500).json(error)
     }
 }
